+++ server/Recruit-Api/src/main/java/com/econovation/recruit/api/applicant/usecase/TimeTableLoadUseCase.java
@@ -2,7 +2,7 @@ package com.econovation.recruit.api.applicant.usecase;
 
 import com.econovation.recruitcommon.annotation.UseCase;
 import com.econovation.recruitdomain.domains.applicant.dto.TimeTableDto;
-import com.econovation.recruitdomain.domains.timetable.TimeTable;
+import com.econovation.recruitdomain.domains.timetable.domain.TimeTable;
 import java.util.List;
 import java.util.Map;
 import java.util.UUID;
