+++ server/Recruit-Api/src/main/java/com/econovation/recruit/api/card/usecase/CardLoadUseCase.java
@@ -2,7 +2,7 @@ package com.econovation.recruit.api.card.usecase;
 
 import com.econovation.recruitcommon.annotation.UseCase;
 import com.econovation.recruitdomain.domains.board.dto.ColumnsResponseDto;
-import com.econovation.recruitdomain.domains.card.Card;
+import com.econovation.recruitdomain.domains.card.domain.Card;
 import com.econovation.recruitdomain.domains.card.dto.CardResponseDto;
 import java.util.List;
 import java.util.Map;
