+++ server/Recruit-Api/src/main/java/com/econovation/recruit/api/score/controller/ScoreController.java
@@ -4,9 +4,7 @@ import static com.econovation.recruitcommon.consts.RecruitStatic.SCORE_SUCCESS_R
 import static com.econovation.recruitcommon.consts.RecruitStatic.SCORE_SUCCESS_UPDATE_MESSAGE;
 
 import com.econovation.recruit.api.score.docs.ScoreExceptionDocs;
-import com.econovation.recruit.application.port.in.RecordUseCase;
 import com.econovation.recruit.application.port.in.ScoreUseCase;
-import com.econovation.recruit.application.utils.EntityMapper;
 import com.econovation.recruitcommon.annotation.ApiErrorExceptionsExample;
 import com.econovation.recruitdomain.domains.dto.CreateScoreDto;
 import com.econovation.recruitdomain.domains.dto.ScoreAverageDto;
@@ -30,6 +28,7 @@ public class ScoreController {
     private final ScoreUseCase scoreUseCase;
 
     @Operation(description = "Score 평가 생성")
+    @ApiErrorExceptionsExample(ScoreExceptionDocs.class)
     @PostMapping("/scores")
     public ResponseEntity<String> createScore(CreateScoreDto createScoreDto) {
         scoreUseCase.createScore(createScoreDto);
@@ -37,6 +36,7 @@ public class ScoreController {
     }
 
     @Operation(description = "Score 평가 수정")
+    @ApiErrorExceptionsExample(ScoreExceptionDocs.class)
     @PutMapping("/scores")
     public ResponseEntity<String> updateScore(CreateScoreDto updateScoreDto) {
         scoreUseCase.updateScore(updateScoreDto);
@@ -51,14 +51,5 @@ public class ScoreController {
         return new ResponseEntity<>(scores, HttpStatus.OK);
     }
 
-    //    @GetMapping("/scores/records")
-    //    public ResponseEntity<Map<String, Object>> getScoresAndRecordsByApplicantId(UUID
-    // applicantId) {
-    //        List<Score> scores = scoreUseCase.getByApplicantId(applicantId);
-    //        Record record = recordUseCase.findByApplicantId(applicantId);
-    //        Map<String, Object> map = new HashMap<>();
-    //        map.put("scores", scores);
-    //        map.put("record", record);
-    //        return new ResponseEntity(map, HttpStatus.OK);
-    //    }
+    // 삭제 기능은 없습니다.
 }
