+++ server/Recruit-Api/src/main/java/com/econovation/recruit/application/utils/EntityMapper.java
@@ -6,8 +6,8 @@ import com.econovation.recruitdomain.domains.dto.*;
 import com.econovation.recruitdomain.domains.interviewer.domain.Interviewer;
 import com.econovation.recruitdomain.domains.interviewer.domain.Role;
 import com.econovation.recruitdomain.domains.record.Record;
-import com.econovation.recruitdomain.domains.score.Score;
-import com.econovation.recruitdomain.domains.timetable.TimeTable;
+import com.econovation.recruitdomain.domains.score.domain.Score;
+import com.econovation.recruitdomain.domains.timetable.domain.TimeTable;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.UUID;
