+++ server/Recruit-Domain/src/main/java/com/econovation/recruitdomain/domains/card/adaptor/CardAdapter.java
@@ -1,7 +1,7 @@
-package com.econovation.recruitdomain.persistence;
+package com.econovation.recruitdomain.domains.card.adaptor;
 
-import com.econovation.recruitdomain.domains.card.Card;
-import com.econovation.recruitdomain.domains.card.CardRepository;
+import com.econovation.recruitdomain.domains.card.domain.Card;
+import com.econovation.recruitdomain.domains.card.domain.CardRepository;
 import com.econovation.recruitdomain.domains.card.exception.CardNotFoundException;
 import com.econovation.recruitdomain.out.CardLoadPort;
 import com.econovation.recruitdomain.out.CardRecordPort;
