+++ server/Recruit-Domain/src/main/java/com/econovation/recruitdomain/domains/card/domain/CardRepository.java
@@ -1,4 +1,4 @@
-package com.econovation.recruitdomain.domains.card;
+package com.econovation.recruitdomain.domains.card.domain;
 
 import java.util.List;
 import java.util.Optional;
