+++ server/Recruit-Domain/src/main/java/com/econovation/recruitdomain/domains/score/domain/Score.java
@@ -1,4 +1,4 @@
-package com.econovation.recruitdomain.domains.score;
+package com.econovation.recruitdomain.domains.score.domain;
 
 import java.util.UUID;
 import javax.persistence.*;
