+++ server/Recruit-Domain/src/main/java/com/econovation/recruitdomain/domains/score/domain/ScoreRepository.java
@@ -1,4 +1,4 @@
-package com.econovation.recruitdomain.domains.score;
+package com.econovation.recruitdomain.domains.score.domain;
 
 import java.util.List;
 import java.util.UUID;
