+++ server/Recruit-Domain/src/main/java/com/econovation/recruitdomain/domains/score/exception/ScoreErrorCode.java
@@ -1,5 +1,8 @@
 package com.econovation.recruitdomain.domains.score.exception;
 
+import static com.econovation.recruitcommon.consts.RecruitStatic.BAD_REQUEST;
+import static com.econovation.recruitcommon.consts.RecruitStatic.NOT_FOUND;
+
 import com.econovation.recruitcommon.annotation.ExplainError;
 import com.econovation.recruitcommon.exception.BaseErrorCode;
 import com.econovation.recruitcommon.exception.ErrorReason;
@@ -7,9 +10,6 @@ import java.lang.reflect.Field;
 import java.util.Objects;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-
-import static com.econovation.recruitcommon.consts.RecruitStatic.BAD_REQUEST;
-import static com.econovation.recruitcommon.consts.RecruitStatic.NOT_FOUND;
 
 @Getter
 @AllArgsConstructor
