+++ server/Recruit-Domain/src/main/java/com/econovation/recruitdomain/domains/score/exception/ScoreInvalidFieldException.java
@@ -1,7 +1,6 @@
 package com.econovation.recruitdomain.domains.score.exception;
 
 import com.econovation.recruitcommon.exception.RecruitCodeException;
-import com.econovation.recruitdomain.domains.label.exception.LabelErrorCode;
 
 public class ScoreInvalidFieldException extends RecruitCodeException {
     public static final ScoreInvalidFieldException EXCEPTION = new ScoreInvalidFieldException();
