+++ server/Recruit-Domain/src/main/java/com/econovation/recruitdomain/domains/timetable/domain/TimeTable.java
@@ -1,4 +1,4 @@
-package com.econovation.recruitdomain.domains.timetable;
+package com.econovation.recruitdomain.domains.timetable.domain;
 
 import java.util.UUID;
 import javax.persistence.*;
